Move PL tests to their own folder.
